+++ crdesigner/ui/gui/view/toolboxes/road_network_toolbox/road_network_toolbox_ui/traffic_light_widget.py
@@ -1,14 +1,13 @@
 from PyQt5.QtCore import *
 from PyQt5.QtGui import *
 from PyQt5.QtWidgets import *
-import math
 
-from commonroad.scenario.traffic_sign import TrafficLightDirection
+from commonroad.scenario.traffic_light import TrafficLightDirection
 
 from crdesigner.ui.gui.utilities.toolbox_ui import CheckableComboBox, PositionButton
 
 
-class TrafficLightWidget():
+class TrafficLightWidget:
     """
     Inherits the traffic light widget setup
     """